Fix ban and kick
As of now, if you mention the bot as a prefix and try to kick or ban someone, it will try to ban the bot too. Furthermore, the ban command would fail everytime that you specified an amount of days afterwards, as .ban takes a number and `days` was a string. Similarly, the `days` variable will now adjust if there is two mentions in the message (bot, to ban/kick), which would previously be `undefined`.