Fix ban and kick (#271)
* hackban and softban

Standard also removed some returns.

* changes

finding the ban from the returned array requires extra code that really isn't necessary.

* Add softban and hackban

* That wasn't needed.

* Fixes for hackban and softban

* even more fixes

add success and error to banMembers

* More additions

* Much more stuff for hackban and softban

 - Support multiple IDs for softban and hackban
 - getREST the user for hackban
 - Small fix to meme
 - Add reason support for hackban and softban

* finishing touches

* standard pls

* Finishing review touches

 - Make softban remove the user's messages up to a day old.

 - Fix variable inconsistency.

 - Fix vague description of softban.

* Fix ban and kick

As of now, if you mention the bot as a prefix and try to kick or ban someone, it will try to ban the bot too. Furthermore, the ban command would fail everytime that you specified an amount of days afterwards, as .ban takes a number and `days` was a string. Similarly, the `days` variable will now adjust if there is two mentions in the message (bot, to ban/kick), which would previously be `undefined`.

* Optimization

 - Removed if/else statements
 - Now `days` will be defined with a regex

* Adhere to requested changes